Firmware boots and can navigate menus

+++ firmware/Sources/Marlin/src/HAL/shared/backtrace/unwarm.cpp
@@ -7,7 +7,7 @@
  * for free and use it as they wish, with or without modifications, and in
  * any context, commercially or otherwise. The only limitation is that I
  * don't guarantee that the software is fit for any purpose or accept any
- * liability for it's use or misuse - this software is without warranty.
+ * liability for its use or misuse - this software is without warranty.
  ***************************************************************************
  * File Description: Utility functions and glue for ARM unwinding sub-modules.
  **************************************************************************/
@@ -79,7 +79,7 @@ void UnwInitState(UnwState * const state,     /**< Pointer to structure to fill.
 
 // Detect if function names are available
 static int __attribute__ ((noinline)) has_function_names() {
-  uint32_t flag_word = ((uint32_t*)(((int32_t)(&has_function_names)) & (-4))) [-1];
+  uint32_t flag_word = ((uint32_t*)(((uint32_t)(&has_function_names)) & (-4))) [-1];
   return ((flag_word & 0xFF000000) == 0xFF000000) ? 1 : 0;
 }
 
@@ -103,8 +103,7 @@ bool UnwReportRetAddr(UnwState * const state, uint32_t addr) {
     // Lets find the function name, if possible
 
     // Align address to 4 bytes
-    //int32_t pf = addr & (-4);
-		uint32_t pf = ((int32_t)addr) & (-4);
+    uint32_t pf = addr & (-4);
 
     // Scan backwards until we find the function name
     uint32_t v;
